[Web] [Refactor] svg 경로 변경 반영 (#39)
