+++ apps/web/src/app/auth/_components/authGithub/authGithub.tsx
@@ -1,5 +1,5 @@
 import Button from '@/shared/components/button/button';
-import Logo from '@/shared/assets/svg/logo_lg.svg';
+import Logo from '@/shared/assets/svgs/logo_lg.svg';
 import {
   bottomParagraphStyle,
   bottonDiv,
